Refactored Localization folders and made (potential) gpf fix
